Fix critical outputLimit bug: uint8_t→float so clamp actually works
outputLimit was stored as uint8_t (limit*100), e.g. 0.15 became 15.
The clamp(value, -15, 15) never triggered since |value| ≤ 1.0,
effectively disabling all actuator limits. Also fix KM_PID_Compute
→ KM_PID_Calculate in tests.

+++ components/km_act/km_act.h
@@ -53,7 +53,7 @@ typedef struct {
     /* DAC (solo accel/freno) */
     uint8_t dacChannel;   // 0 = A, 1 = B
 
-    uint8_t outputLimit; // Range of output allow[0-100]
+    float outputLimit;
     uint32_t lastOutput;
 } ACT_Controller;
